# Conflicts:
#	src/main/java/com/capstone/ar_guideline/services/impl/ARGuidelineServiceImpl.java

+++ src/main/java/com/capstone/ar_guideline/services/IInstructionDetailService.java
@@ -24,5 +24,5 @@ public interface IInstructionDetailService {
 
   Boolean deleteByInstructionId(String instructionId);
 
-  Integer countInstructionDetailByCourseId(String courseId);
+  Long countInstructionDetailByCourseId(String courseId);
 }
